FileExplorerService
First iteration of the file explorer service, this service will be used to iteratively access all files within a project folder, so that they can be read and then turned into embeddings. So far the process of iteratively exploring the directory is done (checking inside every folder and printing the name of the folders and files).

TODO: Read the files
TODO: Make a function which receives an Embbeder and a file so that it can open the file, generate the embeddings for it and return them to later be store. Could also receive the DB and save them directly, still haven't thought about which approach would be better.